+++ NEWS.md
@@ -1,3 +1,13 @@
+## version 1.1.3
+
+---
+
+
+### internal handling of packages
+
+- add renv structure
+
+
 ## version 1.1.2
 
 ---